refactor(time): 네이밍 변경 및 강의 검색 UI 최적화
