+++ apps/time/src/widgets/time-table/ui/LectureSearchContainer.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { LectureSearch } from '.';
+
+export default function LectureSearchContainer() {
+  return (
+    <div className="h-[calc(100vh-202px)] overflow-y-auto p-10">
+      <LectureSearch />
+    </div>
+  );
+}
