+++ apps/time/src/widgets/time-table/ui/index.ts
@@ -1,9 +1,8 @@
 export { default as TimeTable } from './TimeTable';
 export { default as TimeTableFilter } from './TimeTableFilter';
-export { default as TimeTableLayout } from './TimeTableLayout';
 export { default as TimeTableHeader } from './TimeTableHeader';
-export { default as TimeTableModal } from './TimeTableModal';
 export { default as TimeTableContainer } from './TimeTableContainer';
 export { default as TimeTableLectureTable } from './TimeTableLectureTable';
-export { default as TimeTableLectureRemoveModal } from './TimeTableLectureRemoveModal';
 export { default as TimeTableUtilButtons } from './TimeTableUtilButtons';
+export { default as LectureSearchContainer } from './LectureSearchContainer';
+export { default as LectureSearch } from './LectureSearch';
